requirements.txt
pandas
numpy
matplotlib
opencv-python
tqdm
tifffile
devbio-napari
qtpy
natsort
scikit-image
scipy
prettytable
scikit-learn
future
pip
colorama
setuptools
sympy
uncertainties